add namedtuple helper
